+++ src/icemac/ab/calendar/tests/test_install.py
@@ -5,7 +5,7 @@ import icemac.ab.calendar.install
 import icemac.ab.calendar.testing
 import icemac.addressbook.addressbook
 import icemac.addressbook.testing
-import unittest2 as unittest
+import unittest
 
 
 class TestInstall(unittest.TestCase,